Break serialization fixtures out into conftest
